dom/plugins/test/mochitest/test_instance_re-parent.html

<!DOCTYPE html>
<html>
<head>
  <title>Test re-parentinging an instance's DOM node</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="begin()">
  <p id="display"></p>

  <div id="div1">
    <!-- This embed has to have a "src" attribute. Part of this test involves seeing if we
         properly restart plugins that have been added back to a document without a change
         in URL. Not re-loading an object when the URL hasn't changed is a shortcut used for
         some object types. Without a URL, this won't be tested. -->
    <embed id="plugin1" src="loremipsum.txt" type="application/x-test" width="200" height="200"></embed>
  </div>
  <div id="div2">
  </div>

  <script type="application/javascript;version=1.8">
  SimpleTest.waitForExplicitFinish();

  var exceptionThrown = false;
  var p = document.getElementById('plugin1');
  var d1 = document.getElementById('div1');
  var d2 = document.getElementById('div2');

  var destroyed = false;

  function begin() {
    runTests(function() {
      // Callback when finished - set plugin to windowed and repeat the tests

      info("Repeating tests with wmode=window");
      p.setAttribute("wmode", "window");
      d1.appendChild(p);

      // Forces the plugin to be respawned
      p.src = p.src;

      destroyed = false;
      exceptionThrown = false;
      runTests(function () {
        SimpleTest.finish();
      });
    });
  }

  function runTests(callback) {
    // First tests involve moving the instance from one div to another.
    p.startWatchingInstanceCount();
    p.callOnDestroy(function() {
      destroyed = true;
    });

    try {
      d1.removeChild(p);
    } catch (e) {
      exceptionThrown = true;
    }
    is(exceptionThrown, false, "Testing for exception after removeChild.");

    try {
      d2.appendChild(p);
    } catch (e) {
      exceptionThrown = true;
    }
    is(exceptionThrown, false, "Testing for exception after appendChild.");

    is(destroyed, false, "No instances should have been destroyed at this point.");
    is(p.getInstanceCount(), 0, "No new instances should have been created at this point.");

    // Wait for the event loop to spin and ensure the plugin still wasn't touched
    SimpleTest.executeSoon(function() {
      is(destroyed, false, "No instances should have been destroyed at this point.");
      is(p.getInstanceCount(), 0, "No new instances should have been created at this point.");

      // Removing the instance for a full event loop *should* respawn
      d2.removeChild(p);
      SimpleTest.executeSoon(function() {
        d2.appendChild(p);
        SimpleTest.executeSoon(function() {
          try {
            is(p.getInstanceCount(), 1, "One new instance should have been created at this point.");
          } catch (e) {
            exceptionThrown = true;
          }
          is(exceptionThrown, false, "Testing for exception getting instance count from plugin.");

          p.stopWatchingInstanceCount();
          callback.apply(null);
        });
      });
    });
  }
  </script>
</body>
</html>
